requirements.txt
django
pillow
djangorestframework
requests
djangorestframework-simplejwt